+++ firmware/components/driver_sndmixer/sndmixer.h
@@ -5,30 +5,30 @@
 extern "C" {
 #endif
 
-
 /**
  * @brief Structure describing a sound source
  */
 typedef struct {
-	/*! Initialize the sound source. Returns size of data returned per call of fill_buffer. */
-	int (*init_source)(const void *data_start, const void *data_end, int req_sample_rate, void **ctx);
-	/*! Get the actual sample rate at which the source returns data */
-	int (*get_sample_rate)(void *ctx);
-	/*! Decode a bufferful of data. Returns 0 when file ended or something went wrong. Returns amount of bytes in buffer (normally what init_source returned) otherwise. */
-	int (*fill_buffer)(void *ctx, int8_t *buffer);
-	/*! Destroy source, free resources */
-	void (*deinit_source)(void *ctx);
-	/*! Set frequency of synthesizer */
-	void (*set_frequency)(void *ctx, uint16_t frequency);
-	/*! Set waveform of synthesizer */
-	void (*set_waveform)(void *ctx, uint8_t waveform);
+  /*! Initialize the sound source. Returns size of data returned per call of fill_buffer. */
+  int (*init_source)(const void *data_start, const void *data_end, int req_sample_rate, void **ctx);
+  /*! Get the actual sample rate at which the source returns data */
+  int (*get_sample_rate)(void *ctx);
+  /*! Decode a bufferful of data. Returns 0 when file ended or something went wrong. Returns amount
+   * of bytes in buffer (normally what init_source returned) otherwise. */
+  int (*fill_buffer)(void *ctx, int16_t *buffer);
+  /*! Destroy source, free resources */
+  void (*deinit_source)(void *ctx);
+  /*! Set frequency of synthesizer */
+  void (*set_frequency)(void *ctx, uint16_t frequency);
+  /*! Set waveform of synthesizer */
+  void (*set_waveform)(void *ctx, uint8_t waveform);
 } sndmixer_source_t;
 
 /**
  * @brief Initialize the sound mixer
  *
- * @note This function internally calls kchal_sound_start, there is no need to do this in your program
- *       if you use this function to initialize the sound mixer.
+ * @note This function internally calls kchal_sound_start, there is no need to do this in your
+ * program if you use this function to initialize the sound mixer.
  *
  * @param no_channels Amount if sounds to be able to be played simultaneously.
  */
@@ -50,7 +50,8 @@ int sndmixer_queue_wav(const void *wav_start, const void *wav_end, int evictable
 /**
  * @brief Queue the data of a .mod/.xm/.s3m file to be played
  *
- * This queues a piece of tracked music to be played. It will not be actually played until sndmixer_play is called.
+ * This queues a piece of tracked music to be played. It will not be actually played until
+ * sndmixer_play is called.
  *
  * @param mod_start Start of the filedata
  * @param mod_end End of the filedata
@@ -61,7 +62,8 @@ int sndmixer_queue_mod(const void *mod_start, const void *mod_end);
 /**
  * @brief Queue the data of a .mp3 file to be played
  *
- * This queues a piece of mp3 music to be played. It will not be actually played until sndmixer_play is called.
+ * This queues a piece of mp3 music to be played. It will not be actually played until sndmixer_play
+ * is called.
  *
  * @param mp3_start Start of the filedata
  * @param mp3_end End of the filedata
@@ -70,7 +72,7 @@ int sndmixer_queue_mod(const void *mod_start, const void *mod_end);
 int sndmixer_queue_mp3(const void *mp3_start, const void *mp3_end);
 
 typedef ssize_t (*stream_read_type)(void *, void *, size_t);
-int sndmixer_queue_mp3_stream(stream_read_type read_func, void* stream);
+int sndmixer_queue_mp3_stream(stream_read_type read_func, void *stream);
 
 /**
  * @brief Set or unset a sound to looping mode
@@ -95,7 +97,7 @@ void sndmixer_set_volume(int id, int volume);
 
 /**
  * @brief Play a sound
- * 
+ *
  * When a sound is queued, it is not playing yet. Use this call to start playback. You can also
  * use this call to resume a sound paused by sndmixer_pause or sndmixer_pause_all.
  *
@@ -105,7 +107,7 @@ void sndmixer_play(int id);
 
 /**
  * @brief Pause a sound
- * 
+ *
  * Stops playback of the sound. The sound can be resumed with sndmixer_play().
  *
  * @param id ID of the sound, obtained when queueing it
@@ -114,7 +116,7 @@ void sndmixer_pause(int id);
 
 /**
  * @brief Stop a sound, free the sound source and channel it used.
- * 
+ *
  * Stops playback of the sound and frees all associated structures.
  *
  * @param id ID of the sound, obtained when queueing it
@@ -123,9 +125,9 @@ void sndmixer_stop(int id);
 
 /**
  * @brief Pause all playing sounds
- * 
- * This can be used when e.g. the game is paused. Sounds can be individually un-paused afterwards and new sounds
- * can still be queued and played, given enough free/evictable channels.
+ *
+ * This can be used when e.g. the game is paused. Sounds can be individually un-paused afterwards
+ * and new sounds can still be queued and played, given enough free/evictable channels.
  */
 void sndmixer_pause_all();
 
@@ -136,12 +138,10 @@ void sndmixer_pause_all();
  */
 void sndmixer_resume_all();
 
-
-//Basic synthesizer
+// Basic synthesizer
 int sndmixer_queue_synth();
 void sndmixer_freq(int id, uint16_t frequency);
 void sndmixer_waveform(int id, uint8_t waveform);
-
 
 #ifdef __cplusplus
 }
